# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for RTRP application",
        "title": "RTRP API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Kullanici girisi yapmak icin kullanilir. Provider: 0: Guest, 1 Email, 2 Google",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "User login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmetkoprulu_rtrp_models.ApiResponse-models_UserPlayer"
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Email ve Parola ile kayit olmak icin kullanilir. Provider 1 Email anlamina gelir. Identifier alanina email adresi girilir. Secret alanina parola girilir.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration details",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmetkoprulu_rtrp_models.ApiResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_ahmetkoprulu_rtrp_models.ApiResponse-models_UserPlayer": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.UserPlayer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_ahmetkoprulu_rtrp_models.ApiResponse-string": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "internal_api_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error message describing what went wrong"
                }
            }
        },
...
        "models.LoginRequest": {
            "type": "object",
            "required": [
                "identifier"
            ],
            "properties": {
                "identifier": {
                    "description": "email for Email provider, token for social providers",
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/models.SocialNetwork"
                },
                "secret": {
                    "description": "only required for Email provider",
                    "type": "string"
                }
            }
        },
...
        "models.Player": {
            "type": "object",
            "properties": {
                "chips": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "profile_pic_url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.RegisterRequest": {
            "type": "object",
            "required": [
                "identifier",
                "provider"
            ],
            "properties": {
                "identifier": {
                    "description": "email for Email provider, token for social providers",
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/models.SocialNetwork"
                },
                "secret": {
                    "description": "only required for Email provider",
                    "type": "string"
                }
            }
        },
...
        "models.SocialNetwork": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "Guest",
                "Email",
                "Google",
                "Facebook",
                "Apple"
            ]
        },
        "models.UserPlayer": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "player": {
                    "$ref": "#/definitions/models.Player"
                },
                "provider": {
                    "$ref": "#/definitions/models.SocialNetwork"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}